factor out mixed parsers
